requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
sentry-sdk[flask]==1.4.3
yake==0.4.8
torch~=1.10.0
#tensorflow==2.6.0
#transformers[torch]
#transformers[tf-cpu]
bert-extractive-summarizer==0.8.1
deep-translator
pysentimiento==0.2.5
sentencepiece==0.1.96
requests~=2.26.0
numpy~=1.19.5
scipy
blinker
ibm-watson==5.3.0
readtime